corrijo promedio

+++ curso C# nivel1/ejercicios2/NotasPromedio/Program.cs
@@ -7,15 +7,15 @@ namespace NotasPromedio
         static void Main(string[] args)
         {
             //instancio las variables
-            int nota1, nota2, nota3, promedio;
+            float nota1, nota2, nota3, promedio;
 
             //pido por pantalla y asigno el valor a las variables
             Console.WriteLine("Primera Nota: ");
-            nota1 = int.Parse(Console.ReadLine());
+            nota1 = float.Parse(Console.ReadLine());
             Console.WriteLine("Segunda Nota: ");
-            nota2 = int.Parse(Console.ReadLine());
+            nota2 = float.Parse(Console.ReadLine());
             Console.WriteLine("Tercera Nota: ");
-            nota3 = int.Parse(Console.ReadLine());
+            nota3 = float.Parse(Console.ReadLine());
 
             //hago el proceso y lo asigno a una variable
             promedio = (nota1 + nota2 + nota3) / 3;
